requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
openai==1.109.1
pydantic==2.11.9
pydantic_core==2.33.2
python-decouple==3.8
requests==2.31.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
